+++ sympl/__init__.py
@@ -11,14 +11,16 @@ from ._core.exceptions import (
     InvalidPropertyDictError)
 from ._core.array import DataArray
 from ._core.constants import (
-    get_constant, set_constant, set_condensible_name, reset_constants)
+    get_constant, set_constant, set_condensible_name, reset_constants,
+    get_constants_string)
 from ._core.util import (
     combine_array_dimensions,
     ensure_no_shared_keys,
     get_numpy_array, jit,
     restore_dimensions, get_numpy_arrays_with_properties,
     restore_data_arrays_with_properties,
-    set_direction_names, add_direction_names)
+    set_direction_names, add_direction_names,
+    get_component_aliases)
 from ._core.testing import ComponentTestBase
 from ._components import (
     PlotFunctionMonitor, NetCDFMonitor, RestartMonitor,
@@ -26,7 +28,7 @@ from ._components import (
     TimeDifferencingWrapper)
 from ._core.time import datetime, timedelta
 
-__version__ = '0.3.1'
+__version__ = '0.3.2'
 __all__ = (
     Prognostic, Diagnostic, Implicit, Monitor, PrognosticComposite,
     DiagnosticComposite, MonitorComposite, ImplicitPrognostic,
@@ -35,12 +37,12 @@ __all__ = (
     InvalidPropertyDictError,
     DataArray,
     get_constant, set_constant, set_condensible_name, reset_constants,
-    TimeDifferencingWrapper, combine_array_dimensions,
+    get_constants_string,TimeDifferencingWrapper, combine_dimensions,
     ensure_no_shared_keys,
     get_numpy_array, jit,
     restore_dimensions, get_numpy_arrays_with_properties,
     restore_data_arrays_with_properties,
-    set_direction_names, add_direction_names,
+    set_direction_names, add_direction_names, get_component_aliases,
     ComponentTestBase,
     PlotFunctionMonitor, NetCDFMonitor, RestartMonitor,
     ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic,