+++ tests/test_netcdf_monitor.py
@@ -4,6 +4,7 @@ import os
 from datetime import datetime, timedelta
 import numpy as np
 import xarray as xr
+import unittest
 
 random = np.random.RandomState(0)
 
@@ -15,14 +16,103 @@ state = {
     'air_temperature': DataArray(
         random.randn(nx, ny, nz),
         dims=['lon', 'lat', 'mid_levels'],
-        attrs={'units': 'degK'},
+        attrs={'units': 'degK', 'long_name': 'air_temperature'},
     ),
     'air_pressure': DataArray(
         random.randn(nx, ny, nz),
         dims=['lon', 'lat', 'mid_levels'],
-        attrs={'units': 'Pa'},
+        attrs={'units': 'Pa', 'long_name': 'air_pressure'},
     ),
 }
+
+
+class NetCDFMonitorAliasTests(unittest.TestCase):
+
+    def setUp(self):
+        self.ncfile = 'out.nc'
+
+    def tearDown(self):
+        if os.path.isfile(self.ncfile):
+            os.remove(self.ncfile)
+
+    def store_state_and_check_file(self, aliases):
+        assert not os.path.isfile(self.ncfile)
+        monitor = NetCDFMonitor(self.ncfile, aliases=aliases,
+                                write_on_store=True)
+        monitor.store(state)
+        assert os.path.isfile(self.ncfile)
+
+    def check_nc_var(self, varname, varunit, longname):
+        with xr.open_dataset(self.ncfile) as ds:
+            assert len(ds.data_vars.keys()) == 2
+            assert varname in ds.data_vars.keys()
+            assert ds.data_vars[varname].attrs['units'] == varunit
+            assert ds.data_vars[varname].attrs['long_name'] == longname
+            assert tuple(ds.data_vars[varname].shape) == (1, nx, ny, nz)
+
+    def test_state_key_emptystring(self):
+        aliases = {'air_temperature': 'T'}
+        bad_state = {
+                'time': datetime(2013, 7, 20),
+                '': DataArray(
+                    random.randn(nx, ny, nz),
+                    dims=['lon', 'lat', 'mid_levels'],
+                    attrs={'units': 'degK', 'long_name': 'air_temperature'})
+                    }
+        assert not os.path.isfile(self.ncfile)
+        monitor = NetCDFMonitor(self.ncfile, aliases=aliases,
+                                write_on_store=True)
+        self.assertRaises(ValueError, monitor.store, bad_state)
+        assert not os.path.isfile(self.ncfile)
+
+    def test_keys_string_values_string(self):
+        aliases = {'air_temperature': 'T'}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('T', 'degK', 'air_temperature')
+        self.check_nc_var('air_pressure', 'Pa', 'air_pressure')
+
+    def test_keys_nonstring_values_string(self):
+        aliases = {1.0: 'T'}
+        assert not os.path.isfile(self.ncfile)
+        self.assertRaises(TypeError, NetCDFMonitor, self.ncfile,
+                          aliases=aliases, write_on_store=True)
+        assert not os.path.isfile(self.ncfile)
+
+    def test_keys_string_values_nonstring(self):
+        aliases = {'air_temperature': 1.0}
+        assert not os.path.isfile(self.ncfile)
+        self.assertRaises(TypeError, NetCDFMonitor, self.ncfile,
+                          aliases=aliases, write_on_store=True)
+        assert not os.path.isfile(self.ncfile)
+
+    def test_keys_string_values_emptystring(self):
+        # this SHOULD raise a ValueError
+        aliases = {'air_temperature': ''}
+        assert not os.path.isfile(self.ncfile)
+        monitor = NetCDFMonitor(self.ncfile, aliases=aliases,
+                                write_on_store=True)
+        self.assertRaises(ValueError, monitor.store, state)
+        assert not os.path.isfile(self.ncfile)
+
+    def test_keys_partialstring_values_emptystring(self):
+        # this SHOULD NOT raise a ValueError
+        aliases = {'air_': ''}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('temperature', 'degK', 'air_temperature')
+        self.check_nc_var('pressure', 'Pa', 'air_pressure')
+
+    def test_empty_aliases(self):
+        aliases = {}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('air_temperature', 'degK', 'air_temperature')
+        self.check_nc_var('air_pressure', 'Pa', 'air_pressure')
+
+    def test_two_aliases(self):
+        aliases = {'air_temperature': 'T',
+                   'air_pressure': 'P'}
+        self.store_state_and_check_file(aliases)
+        self.check_nc_var('T', 'degK', 'air_temperature')
+        self.check_nc_var('P', 'Pa', 'air_pressure')
 
 
 def test_netcdf_monitor_initializes():