+++ HISTORY.rst
@@ -45,6 +45,23 @@ Breaking changes
   for calls to Diagnostic, Prognostic, ImplicitPrognostic, and Implicit components,
   and may be strictly required in other ways in the future
 
+v0.3.2
+------
+
+* Exported get_constants_string to the public API
+* Added "aliases" kwarg to NetCDFMonitor, allowing the monitor to shorten
+  variable names when writing to netCDF
+* Added get_component_aliases() to get a dictionary of quantity aliases from
+  a list of Components (used by NetCDFMonitor to shorten variable
+  names)
+* Added tests for NetCDFMonitor aliases and get_component_aliases()
+
+Breaking changes
+~~~~~~~~~~~~~~~~
+* tendencies in diagnostics are now named as X_tendency_from_Y, instead of
+  tendency_of_X_due_to_Y. The idea is that it's shorter, and can easily be
+  shortened more by aliasing "tendency" to "tend"
+
 v0.3.1
 ------